requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
Brotli==1.0.9
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dash==1.16.3
dash-bootstrap-components==0.10.7
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
decorator==4.4.2
Flask==1.1.2
Flask-Compress==1.7.0
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.34.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
numba==0.51.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
Pillow==8.0.0
plotly==4.12.0
pooch==1.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
pynndescent==0.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.3
seaborn==0.11.0
six==1.15.0
sounddevice==0.4.1
SoundFile==0.10.3.post1
tbb==2020.3.254
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
umap-learn==0.5.0
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
